# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Stade de Mbour",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/clubs": {
            "get": {
                "tags": [
                    "Clubs"
                ],
                "summary": "Lister tous les clubs",
                "operationId": "c44731a8866e74bcd14a6b4f0f1f82b6",
                "responses": {
                    "200": {
                        "description": "Liste des clubs récupérée avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Club"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Non autorisé"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Clubs"
                ],
                "summary": "Créer un nouveau club",
                "operationId": "4e37feabfe957444c16ab299c31af0a7",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "AS Dakar"
                                    },
                                    "date_creation": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2020-01-01"
                                    },
                                    "stade": {
                                        "type": "string",
                                        "example": "Stade Léopold Sédar Senghor"
                                    },
                                    "entraineur": {
                                        "type": "string",
                                        "example": "Jean Dupont"
                                    },
                                    "president": {
                                        "type": "string",
                                        "example": "Pierre Ndiaye"
                                    },
                                    "user_id": {
                                        "type": "string",
                                        "format": "uuid",
                                        "example": "a2f9c8c1-72ab-4a6b-9d74-3f4078ce7e99"
                                    },
                                    "logo": {
                                        "description": "Logo du club",
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Club créé avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Club"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Erreur de validation"
                    },
                    "401": {
                        "description": "Non autorisé"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/clubs/{id}": {
            "get": {
                "tags": [
                    "Clubs"
                ],
                "summary": "Afficher les détails d'un club",
                "operationId": "b5900ff997f9d991c3a388875d8971f8",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "UUID du club",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Détails du club récupérés avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Club"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Club non trouvé"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Clubs"
                ],
                "summary": "Mettre à jour un club existant",
                "operationId": "9a7201b06d26cfc6aa54fe93f98774d2",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "UUID du club",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "AS Dakar"
                                    },
                                    "date_creation": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2020-01-01"
                                    },
                                    "stade": {
                                        "type": "string",
                                        "example": "Stade Léopold Sédar Senghor"
                                    },
                                    "entraineur": {
                                        "type": "string",
                                        "example": "Jean Dupont"
                                    },
                                    "president": {
                                        "type": "string",
                                        "example": "Pierre Ndiaye"
                                    },
                                    "user_id": {
                                        "type": "string",
                                        "format": "uuid"
                                    },
                                    "logo": {
                                        "description": "Logo du club",
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Club mis à jour avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Club"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Club non trouvé"
                    },
                    "401": {
                        "description": "Non autorisé"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Clubs"
                ],
                "summary": "Supprimer un club",
                "operationId": "7ae9561f331895c99ba0d047eb68ba96",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "UUID du club",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Club supprimé avec succès"
                    },
                    "404": {
                        "description": "Club non trouvé"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "Club": {
                "title": "Club",
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid",
                        "example": "550e8400-e29b-41d4-a716-446655440000"
                    },
                    "name": {
                        "type": "string",
                        "example": "AS Dakar"
                    },
                    "date_creation": {
                        "type": "string",
                        "format": "date",
                        "example": "2020-01-01"
                    },
                    "stade": {
                        "type": "string",
                        "example": "Stade Léopold Sédar Senghor"
                    },
                    "entraineur": {
                        "type": "string",
                        "example": "Jean Dupont"
                    },
                    "president": {
                        "type": "string",
                        "example": "Pierre Ndiaye"
                    },
                    "user_id": {
                        "type": "string",
                        "format": "uuid",
                        "example": "f3f9c8c1-72ab-4a6b-9d74-3f4078ce7e84"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
    "security": [
        {
            "bearerAuth": []
        }
    ]
}